requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
PyQt5==5.15.10 